refactor: separate folder for test file
