+++ tests/Feature/Admin/ContentTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Admin;
 
 use App\Enum\Contents\ContentType;
 use App\Enum\Users\RoleEnum;
-use App\Models\User;
 use App\Models\Content;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
